# omod/src/main/webapp/resources/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "REST API for Biometric Module.",
    "version" : "v1",
    "title" : "Biometric Module REST API"
  },
  "basePath" : "/openmrs/ws",
...
  "paths" : {
...
    "/rest/v1/biometric/devicename" : {
      "post" : {
        "tags" : [ "REST API for accessing User Information and saving operator device" ],
        "summary" : "Save readable device name",
        "description" : "Save readable device name",
        "operationId" : "saveReadableDeviceName",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "header",
          "required" : true,
          "type" : "string",
          "default" : ""
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "location details",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "On successful saving of device name",
            "schema" : {
              "$ref" : "#/definitions/DeviceNameResponse"
            }
          },
          "400" : {
            "description" : "Error in request details shared"
          },
          "500" : {
            "description" : "Failure to save device name"
          }
        }
      }
    },
...
    "/rest/v1/biometric/users" : {
      "get" : {
        "tags" : [ "REST API for accessing User Information and saving operator device" ],
        "summary" : "Get All User Details",
        "description" : "Get All User Details",
        "operationId" : "getAllUsers",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "On successful return of all user details",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "object"
              }
            }
          },
          "500" : {
            "description" : "Failure to return user details"
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "DeviceNameResponse" : {
      "type" : "object",
      "properties" : {
        "deviceName" : {
          "type" : "string"
        }
      }
    },
...
  }
}